Fixing integration failures (#439)
## 📝 Description

**What does this PR do and why is this change necessary?**

## ✔️ How to Test

**What are the steps to reproduce the issue or verify the changes?**

**How do I run the relevant unit/integration tests?**

## 📷 Preview

**If applicable, include a screenshot or code snippet of this change.
Otherwise, please remove this section.**
