+++ tests/integration/linodes/helpers_linodes.py
@@ -1,9 +1,7 @@
 import os
 import time
 
-import pytest
-
-from tests.integration.helpers import delete_target_id, exec_test_command
+from tests.integration.helpers import exec_test_command
 
 DEFAULT_RANDOM_PASS = (
     exec_test_command(["openssl", "rand", "-base64", "32"])
@@ -205,195 +203,3 @@ def create_linode_and_wait(
     )
 
     return linode_id
-
-
-@pytest.fixture
-def create_linode_with_label():
-    result = (
-        exec_test_command(
-            BASE_CMD
-            + [
-                "create",
-                "--type",
-                "g6-standard-2",
-                "--region",
-                "us-east",
-                "--image",
-                DEFAULT_TEST_IMAGE,
-                "--label",
-                "cli-1",
-                "--root_pass",
-                DEFAULT_RANDOM_PASS,
-                "--text",
-                "--delimiter",
-                ",",
-                "--no-headers",
-                "--format",
-                "label,region,type,image,id",
-                "--no-defaults",
-            ]
-        )
-        .stdout.decode()
-        .rstrip()
-    )
-
-    yield result
-
-    res_arr = result.split(",")
-    linode_id = res_arr[4]
-    delete_target_id(target="linodes", id=linode_id)
-
-
-@pytest.fixture
-def create_linode_min_req():
-    result = (
-        exec_test_command(
-            BASE_CMD
-            + [
-                "create",
-                "--type",
-                "g6-standard-2",
-                "--region",
-                "us-east",
-                "--root_pass",
-                DEFAULT_RANDOM_PASS,
-                "--no-defaults",
-                "--text",
-                "--delimiter",
-                ",",
-                "--no-headers",
-                "--format",
-                "id,region,type",
-            ]
-        )
-        .stdout.decode()
-        .rstrip()
-    )
-
-    yield result
-
-    res_arr = result.split(",")
-    linode_id = res_arr[0]
-    delete_target_id(target="linodes", id=linode_id)
-
-
-@pytest.fixture
-def create_linode_wo_image():
-    linode_type = (
-        os.popen(
-            "linode-cli linodes types --text --no-headers --format=id | xargs | awk '{ print $1 }'"
-        )
-        .read()
-        .rstrip()
-    )
-    linode_region = (
-        os.popen(
-            "linode-cli regions list --format=id  --text --no-headers | xargs | awk '{ print $1 }'"
-        )
-        .read()
-        .rstrip()
-    )
-
-    exec_test_command(
-        BASE_CMD
-        + [
-            "create",
-            "--no-defaults",
-            "--label",
-            "cli-2",
-            "--type",
-            linode_type,
-            "--region",
-            linode_region,
-            "--root_pass",
-            DEFAULT_RANDOM_PASS,
-        ]
-    ).stdout.decode()
-
-    linode_id = (
-        exec_test_command(
-            BASE_CMD
-            + [
-                "list",
-                "--label",
-                "cli-2",
-                "--text",
-                "--no-headers",
-                "--format",
-                "id",
-            ]
-        )
-        .stdout.decode()
-        .rstrip()
-    )
-
-    yield linode_id
-
-    delete_target_id(target="linodes", id=linode_id)
-
-
-@pytest.fixture
-def create_linode_backup_enabled():
-    linode_type = (
-        os.popen(
-            "linode-cli linodes types --text --no-headers --format='id' | xargs | awk '{ print $1 }'"
-        )
-        .read()
-        .rstrip()
-    )
-
-    # create linode with backups enabled
-    linode_id = (
-        exec_test_command(
-            [
-                "linode-cli",
-                "linodes",
-                "create",
-                "--backups_enabled",
-                "true",
-                "--type",
-                linode_type,
-                "--region",
-                DEFAULT_REGION,
-                "--image",
-                DEFAULT_TEST_IMAGE,
-                "--root_pass",
-                DEFAULT_RANDOM_PASS,
-                "--text",
-                "--no-headers",
-                "--format=id",
-            ]
-        )
-        .stdout.decode()
-        .rstrip()
-    )
-
-    yield linode_id
-
-    delete_target_id("linodes", linode_id)
-
-
-@pytest.fixture
-def take_snapshot_of_linode():
-    timestamp = str(time.time())
-    # get linode id after creation and wait for "running" status
-    linode_id = create_linode_and_wait()
-    new_snapshot_label = "test_snapshot" + timestamp
-
-    result = exec_test_command(
-        BASE_CMD
-        + [
-            "snapshot",
-            linode_id,
-            "--label",
-            new_snapshot_label,
-            "--text",
-            "--delimiter",
-            ",",
-            "--no-headers",
-        ]
-    ).stdout.decode()
-
-    yield linode_id, new_snapshot_label
-
-    delete_target_id("linodes", linode_id)
